+++ db/errors.go
@@ -1,5 +1,0 @@
-package db
-
-import "errors"
-
-var ErrMigrationsPathEmpty = errors.New("migrations path is empty")