+++ db/postgresql/migrate.go
@@ -2,102 +2,40 @@ package postgresql
 
 import (
 	"embed"
-	"errors"
-	"fmt"
-	"log/slog"
-	"os"
 
-	"github.com/golang-migrate/migrate/v4"
+	"github.com/golang-migrate/migrate/v4/database"
 	"github.com/golang-migrate/migrate/v4/database/pgx/v5"
-	_ "github.com/golang-migrate/migrate/v4/source/file" // needed for file:// scheme
-	"github.com/golang-migrate/migrate/v4/source/iofs"
+	"github.com/psyb0t/common-go/db"
 	"github.com/psyb0t/ctxerrors"
 )
 
-func (p *Postgresql) createMigrator(
-	path string,
-	fs *embed.FS,
-) (*migrate.Migrate, error) {
+func (p *Postgresql) migrationDriver() (database.Driver, error) {
 	driver, err := pgx.WithInstance(p.SQLDB, &pgx.Config{})
 	if err != nil {
 		return nil, ctxerrors.Wrap(err, "could not create database driver")
 	}
 
-	if fs != nil {
-		sourceDriver, err := iofs.New(fs, path)
-		if err != nil {
-			return nil, ctxerrors.Wrap(err, "could not create iofs driver")
-		}
-
-		m, err := migrate.NewWithInstance(
-			"iofs",
-			sourceDriver,
-			p.config.Database,
-			driver,
-		)
-		if err != nil {
-			return nil, ctxerrors.Wrap(err, "could not create migrate instance")
-		}
-
-		return m, nil
-	}
-
-	if path == "" {
-		return nil, ctxerrors.Wrap(ErrMigrationsPathRequired, "path is empty")
-	}
-
-	if _, err := os.Stat(path); err != nil {
-		return nil, ctxerrors.Wrap(err, "error validating migrations path")
-	}
-
-	m, err := migrate.NewWithDatabaseInstance(
-		fmt.Sprintf("file://%s", path),
-		p.config.Database,
-		driver,
-	)
-	if err != nil {
-		return nil, ctxerrors.Wrap(err, "could not create migrate instance")
-	}
-
-	return m, nil
+	return driver, nil
 }
 
 // MigrateUp applies all available migrations.
 func (p *Postgresql) MigrateUp(path string, fs *embed.FS) error {
-	m, err := p.createMigrator(path, fs)
+	driver, err := p.migrationDriver()
 	if err != nil {
-		return ctxerrors.Wrap(err, "could not create migrator")
+		return ctxerrors.Wrap(err, "create PostgreSQL migration driver")
 	}
 
-	if err := m.Up(); err != nil {
-		if errors.Is(err, migrate.ErrNoChange) {
-			slog.Info("database is already up to date")
-
-			return nil
-		}
-
-		return ctxerrors.Wrap(err, "could not migrate up")
-	}
-
-	return nil
+	return db.MigrateUp(p.config.Database, driver, path, fs)
 }
 
 // MigrateDown reverts the specified number of migrations.
 func (p *Postgresql) MigrateDown(path string, steps int, fs *embed.FS) error {
-	if steps <= 0 {
-		return ctxerrors.Wrap(ErrInvalidSteps, "steps must be greater than 0")
-	}
-
-	m, err := p.createMigrator(path, fs)
+	driver, err := p.migrationDriver()
 	if err != nil {
-		return ctxerrors.Wrap(err, "could not create migrator")
+		return ctxerrors.Wrap(err, "create PostgreSQL migration driver")
 	}
 
-	if err := m.Steps(-steps); err != nil {
-		return ctxerrors.Wrap(err, "could not migrate down")
-	}
-
-	return nil
+	return db.MigrateDown(p.config.Database, driver, path, fs, steps)
 }
 
 // MigrateForce forces migration to a specific version.
@@ -106,14 +44,10 @@ func (p *Postgresql) MigrateForce(
 	version int,
 	fs *embed.FS,
 ) error {
-	m, err := p.createMigrator(path, fs)
+	driver, err := p.migrationDriver()
 	if err != nil {
-		return ctxerrors.Wrap(err, "could not create migrator")
+		return ctxerrors.Wrap(err, "create PostgreSQL migration driver")
 	}
 
-	if err := m.Force(version); err != nil {
-		return ctxerrors.Wrap(err, "could not force version")
-	}
-
-	return nil
+	return db.MigrateForce(p.config.Database, driver, path, fs, version)
 }